Fix warnings found in DEBUG combined build
Fixes:
* Irrelevant where conditions
* Irrelevant const
* A potential infinite loop in treenode
* Bad implicit case fallthroughs
* Explicit markings for required case fallthroughs
* Unused variables
* Unused function

Also disabled some warnings for now which we should fix later.

+++ dbcon/execplan/treenode.h
@@ -187,7 +187,7 @@ typedef SP_IDB_Regex SP_CNX_Regex;
 inline std::string removeTrailing0(char* val, uint32_t length)
 {
     char* ptr = val;
-    uint32_t i = 0;
+    int64_t i = 0;
     bool decimal_point = false;
 
     for (; i < length; i++, ptr++)
